Fix SSR Theme Context (#11)
