+++ src/contexts/theme.js
@@ -1,25 +1,11 @@
 import React from "react";
-import { STORAGE_VALUE } from "../constants/storageValue";
-import { THEME_MODE } from "../constants/theme";
 import { useThemeMode } from "../hooks/useThemeMode";
 
 const ThemeContext = React.createContext({});
 
 // eslint-disable-next-line react/prop-types
 const ThemeContextProvider = ({ children }) => {
-  // Check User Preffered Media Color
-  const { matches: prefferedUserColor } = window.matchMedia(
-    `(prefers-color-scheme: light)`
-  );
-  const hasPredefinedColor = localStorage.getItem(STORAGE_VALUE.THEME);
-  const prefferedUserMediaColor = prefferedUserColor
-    ? THEME_MODE.LIGHT
-    : THEME_MODE.DARK;
-  // Assign Theme Mode based on User Prefferend Media Color
-  const defaultThemeValue =
-    hasPredefinedColor !== null ? hasPredefinedColor : prefferedUserMediaColor;
-
-  const value = useThemeMode({ defaultTheme: defaultThemeValue });
+  const value = useThemeMode();
   return (
     <ThemeContext.Provider value={{ ...value }}>
       {children}
